新增: 在初步设计阶段未完成的所有接口（不包括废弃的） 修复: 定时任务未全局启用的问题 删除：删除了 indicator 统计指标接口 修改：修改了部分接口的URL，修复一些拼写错误

+++ src/main/kotlin/top/goopper/platform/controller/course/SectionController.kt
@@ -2,14 +2,7 @@ package top.goopper.platform.controller.course
 
 import org.springframework.http.ResponseEntity
 import org.springframework.security.access.prepost.PreAuthorize
-import org.springframework.web.bind.annotation.DeleteMapping
-import org.springframework.web.bind.annotation.GetMapping
-import org.springframework.web.bind.annotation.PathVariable
-import org.springframework.web.bind.annotation.PostMapping
-import org.springframework.web.bind.annotation.PutMapping
-import org.springframework.web.bind.annotation.RequestBody
-import org.springframework.web.bind.annotation.RequestMapping
-import org.springframework.web.bind.annotation.RestController
+import org.springframework.web.bind.annotation.*
 import top.goopper.platform.dto.course.create.CreateSectionDTO
 import top.goopper.platform.pojo.Response
 import top.goopper.platform.service.course.SectionService
@@ -24,7 +17,7 @@ class SectionController(
     @PostMapping("/{courseId}")
     fun createNewSection(
         @RequestBody section: CreateSectionDTO,
-        @PathVariable courseId: Long
+        @PathVariable courseId: Int
     ): ResponseEntity<Response> {
         section.courseId = courseId
         sectionService.createNewSection(section)
@@ -34,7 +27,7 @@ class SectionController(
     @PreAuthorize("hasAnyRole('ADMIN', 'TEACHER')")
     @GetMapping("/creation/{sectionId}")
     fun getCreationInfo(
-        @PathVariable sectionId: Long
+        @PathVariable sectionId: Int
     ): ResponseEntity<Response> {
         val creation = sectionService.getCreationInfo(sectionId)
         return ResponseEntity.ok(Response.success(creation))
@@ -52,15 +45,18 @@ class SectionController(
     @PreAuthorize("hasAnyRole('ADMIN', 'TEACHER')")
     @DeleteMapping("/{sectionId}")
     fun deleteSection(
-        @PathVariable sectionId: Long
+        @PathVariable sectionId: Int
     ): ResponseEntity<Response> {
         sectionService.deleteSection(sectionId)
         return ResponseEntity.ok(Response.success())
     }
 
+    /**
+     * 获取章节详细信息
+     */
     @GetMapping("/{sectionId}")
     fun getSectionDetail(
-        @PathVariable sectionId: Long
+        @PathVariable sectionId: Int
     ): ResponseEntity<Response> {
         val section = sectionService.getSectionDetail(sectionId)
         return ResponseEntity.ok(Response.success(section))
